consolidate theme files. Nothing to see here
